# api-gateway/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  booking.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  booking.ListProvidersResponse:
    properties:
      providers:
        description: List of providers
        items:
          $ref: '#/definitions/booking.ProviderResponse'
        type: array
    type: object
...
  booking.ProviderResponse:
    properties:
      average_rating:
        type: number
      company_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/booking.GeoPoint'
      services:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  booking.UpdateProviderRequest:
    properties:
      availability:
        description: Updated array of available time slots
        type: string
      average_rating:
        description: Updated average rating
        type: number
      company_name:
        description: Updated company name
        type: string
      description:
        description: Updated description
        type: string
      id:
        description: Provider ID
        type: string
      location:
        allOf:
        - $ref: '#/definitions/booking.GeoPoint'
        description: Updated location
      services:
        description: Updated array of service IDs
        items:
          type: string
        type: array
    type: object
...
  models.GeoPoint:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  models.Provider:
    properties:
      availability:
        type: string
      average_rating:
        type: number
      company_name:
        type: string
      description:
        type: string
      location:
        $ref: '#/definitions/models.GeoPoint'
      services:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway of On-Demand Car Wash Service
  title: On-Demand Car Wash Service
  version: "1.0"
paths:
...
  /admin/providers/{id}:
    get:
      description: Retrieves provider details by provider ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderResponse'
        "400":
          description: Invalid provider ID
          schema:
            type: string
        "404":
          description: Provider not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a provider by ID
      tags:
      - providers
...
  /customer/providers:
    get:
      description: Retrieves a list of providers with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Page limit
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListProvidersResponse'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Lists all providers with pagination
      tags:
      - providers
  /customer/providers/{id}:
    get:
      description: Retrieves provider details by provider ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderResponse'
        "400":
          description: Invalid provider ID
          schema:
            type: string
        "404":
          description: Provider not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a provider by ID
      tags:
      - providers
...
  /provider:
    get:
      description: Retrieves a list of providers with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Page limit
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ListProvidersResponse'
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Lists all providers with pagination
      tags:
      - providers
    post:
      description: Registers a new provider and returns the provider details
      parameters:
      - description: Provider details
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/models.Provider'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderResponse'
        "400":
          description: Invalid provider data
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Registers a new provider
      tags:
      - providers
  /provider/{id}:
    delete:
      description: Deletes a provider by provider ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid provider ID
          schema:
            type: string
        "404":
          description: Provider not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Deletes a provider by ID
      tags:
      - providers
    get:
      description: Retrieves provider details by provider ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderResponse'
        "400":
          description: Invalid provider ID
          schema:
            type: string
        "404":
          description: Provider not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieves a provider by ID
      tags:
      - providers
    put:
      description: Updates provider details by provider ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated provider details
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/booking.UpdateProviderRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderResponse'
        "400":
          description: Invalid provider data or ID
          schema:
            type: string
        "404":
          description: Provider not found
          schema:
            type: string
        "500":
          description: Server error while processing request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates a provider by ID
      tags:
      - providers
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"